Se Completan Endpoints y se edita archivo readme

+++ Back-End/models/order.model.js
@@ -7,8 +7,8 @@ const orderModel = sequelize.define(
 	'order',
 	{
 		quantity: {
-			type: DataTypes.INTEGER,
-			allowNull: false,
+			type: DataTypes.INTEGER(10),
+			allowNull: false
 		},
 	},
 	{
